Split the project into submodules and add EventStoreManager from the same organization
