+++ domain/src/main/java/fr/esir/jxc/domain/models/Address.java
@@ -1,0 +1,18 @@
+package fr.esir.jxc.domain.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Address {
+
+    private String postalCode;
+    private String street;
+    private int streetNumber;
+    private String complement;
+
+}
+
