+++ domain/src/main/java/fr/esir/jxc/domain/models/Article.java
@@ -1,0 +1,25 @@
+package fr.esir.jxc.domain.models;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@RequiredArgsConstructor
+public class Article {
+
+    @NonNull
+    private String id;
+    @NonNull
+    private String url;
+
+    private List<String> tags;
+    private List<String> suggestedTags;
+
+}
