+++ domain/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+server.port=8080
+
+url.base = "/api/v1"
+url.paths.users = "/users"
