+++ elastic-search/src/test/java/fr/esir/jxc/domain/ElasticSearchApplicationTests.java
@@ -1,0 +1,16 @@
+package fr.esir.jxc.domain;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ElasticSearchApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
