+++ kafka/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+kafka:
+    topic: test
+    bootstrapAddress: localhost:9092
+    groupId: fr.esir.jxc
