--- src/main/java/com/jxc/dbmanager/models/Address.java
@@ -1,18 +1,0 @@
-package com.jxc.dbmanager.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Address {
-
-    private String postalCode;
-    private String street;
-    private int streetNumber;
-    private String complement;
-
-}
-
