--- src/main/java/com/jxc/dbmanager/models/Article.java
@@ -1,21 +1,0 @@
-package com.jxc.dbmanager.models;
-
-import lombok.*;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class Article {
-
-    @NonNull
-    private String id;
-    @NonNull
-    private String url;
-
-    private List<String> tags;
-    private List<String> suggestedTags;
-
-}
