--- src/test/java/com/jxc/dbmanager/DbmanagerApplicationTests.java
@@ -1,16 +1,0 @@
-package com.jxc.dbmanager;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DbmanagerApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
